backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.2
streamlit==1.25.0
requests==2.31.0
pydantic==1.10.13
python-dotenv==1.0.1
google-api-python-client==2.87.0
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-generativeai==0.4.1
langgraph==0.0.11
dateparser==1.1.8
tzlocal==4.2
pytz==2023.3